evals/BIER/requirements.txt
# BEIR evaluation framework
beir>=2.0.0

# Scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Evaluation metrics
pytrec-eval>=0.5
ranx>=0.3.0

# Progress tracking
tqdm>=4.62.0

# Configuration
PyYAML>=6.0

# Optional: Additional IR metrics
ir-measures>=0.3.0

# Compatibility with existing pipeline
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0

# Logging and utilities
loguru>=0.6.0
click>=8.0.0